refactor: fix transform animation exception
